[NFC] Move DroppedVariableStats code to Analysis
